Extract some common MG transfer setup code
