+++ source/multigrid/mg_transfer_internal.inst.in
@@ -35,3 +35,55 @@ for (deal_II_dimension : DIMENSIONS;
     \}
     \}
 }
+
+
+for (deal_II_dimension : DIMENSIONS)
+{
+    namespace internal
+    \{
+    namespace MGTransfer
+    \{
+
+    template
+    unsigned int
+    compute_shift_within_children<deal_II_dimension>(const unsigned int,
+            const unsigned int,
+            const unsigned int);
+    \}
+    \}
+}
+
+for (S : REAL_SCALARS)
+{
+    namespace internal
+    \{
+    namespace MGTransfer
+    \{
+
+    template
+    struct ElementInfo<S>;
+    \}
+    \}
+}
+
+for (deal_II_dimension : DIMENSIONS; S : REAL_SCALARS)
+{
+    namespace internal
+    \{
+    namespace MGTransfer
+    \{
+
+    template
+    void setup_transfer<deal_II_dimension>(const dealii::DoFHandler<deal_II_dimension>&,
+                                           const MGConstrainedDoFs*,
+                                           ElementInfo<S>&,
+                                           std::vector<std::vector<unsigned int> >&,
+                                           std::vector<std::vector<std::pair<unsigned int,unsigned int> > >&,
+                                           std::vector<unsigned int>&,
+                                           std::vector<std::vector<std::vector<unsigned short> > >&,
+                                           std::vector<AlignedVector<VectorizedArray<S> > >&,
+                                           std::vector<std::vector<std::pair<unsigned int, unsigned int> > >&,
+                                           MGLevelObject<LinearAlgebra::distributed::Vector<S> >&);
+    \}
+    \}
+}
